backend/requirements-gcp.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1
pgvector==0.2.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Google Cloud
google-cloud-aiplatform==1.40.0
vertexai==0.0.1

# Data Processing
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3

# Utilities
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0
python-slugify==8.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
